github: add action to run integration tests
Add action to run integration tests.
