+++ tests/test-td-paging/Cargo.toml
@@ -15,6 +15,7 @@ bootloader = "=0.10.9"
 td-paging = { path = "../../td-paging" }
 td-layout = { path = "../../td-layout" }
 test-runner-client = { path = "../../devtools/test-runner-client" }
+x86_64 = "=0.14.6"
 
 [package.metadata.bootloader]
 map-physical-memory = true
